Assess PR interface cleanup
