+++ src/models/mod.rs
@@ -1,4 +1,3 @@
-#![allow(unused_imports)]
 #![allow(ambiguous_glob_reexports)] // We have several test modules, should be fine
 pub mod class;
 pub mod group;
