+++ src/models/traits/mod.rs
@@ -6,9 +6,13 @@ pub mod object_relation;
 pub mod output;
 pub mod user;
 
+#[allow(unused_imports)]
 pub use crate::models::traits::class::*;
+#[allow(unused_imports)]
 pub use crate::models::traits::class_relation::*;
+#[allow(unused_imports)]
 pub use crate::models::traits::namespace::*;
+#[allow(unused_imports)]
 pub use crate::models::traits::object::*;
 pub use crate::models::traits::object_relation::*;
 pub use crate::models::traits::output::*;
